backend/requirements copy.txt
azure-storage-blob
django
djangorestframework
gunicorn
dj-rest-auth
pytest-dotenv
pandas
psycopg2-binary
unipath>=1.0
django-cors-headers
django-filter
numpy
scipy
pytz
celery>=5.2.0,<5.3
django-celery-beat>=2.3.0
django-redis>=5.2.0
channels>=3.0.0,<3.1
channels-redis>=3.4.0,<3.5
scikit-learn
djangorestframework-simplejwt>=5.2.0
pyjwt>=2.6.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
django-storages>=1.13.1
azure-storage-blob>=12.0.0
django-cleanup>=6.0.0
regex==2022.10.31
sorl-thumbnail>=12.9.0
Pillow==9.2.0
sentry-sdk>=1.11.1
python-magic>=0.4.27
mypy>=0.991
django-stubs>=1.13.1
djangorestframework-stubs>=1.8.0
types-docutils==0.19.1.1
types-Markdown==3.4.2.1
types-psycopg2==2.9.21.2
types-Pygments==2.13.1.1
types-python-dateutil==2.8.19.5
types-pytz==2022.7.0.0
types-PyYAML==6.0.12.2
types-requests==2.28.11.7
types-setuptools==65.6.0.2
types-urllib3==1.26.25.4
types-regex==2022.10.31.6
deepdiff>=6.2.2
flake8>=5.0.4
kombu==5.3.0b1
azure-servicebus>=7.8.2
pydantic>=1.10.4